Switches from syscall to golang.org/x/sys/unix
According to
https://golang.org/pkg/syscall/
and
https://docs.google.com/document/d/1QXzI9I1pOfZPujQzxhyRy6EeHYTQitKKjHfpq0zpxZs/edit
the syscall library is essentially deprecated
moving forward with go 1.4.

"NOTE: This package is locked down. Code outside
the standard Go repository should be migrated to
use the corresponding package in the go.sys
subrepository. That is also where updates
required by new systems or versions should be
applied."